requirements.txt
azure-core==1.25.0
azure-functions==1.11.2
azure-storage-blob==12.13.1
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==37.0.4
idna==3.3
isodate==0.6.1
msrest==0.7.1
oauthlib==3.2.0
packaging==21.3
pip-system-certs==3.0
pycparser==2.21
pyparsing==3.0.9
requests==2.28.1
requests-oauthlib==1.3.1
setuptools-scm==7.0.5
six==1.16.0
tomli==2.0.1
typing_extensions==4.3.0
urllib3==1.26.12
wrapt==1.14.1
